feat: mature restructure of proprietary library loading
This gives us the ability to download the required libraries from the web and then loading them.

It makes everything a bit more complicated, as we don't use the #[link] parts anymore, but we don't need it anyways if this works as well.

This also removes our custom linking and all files related to it.